Move renderer to separate file renderer.py
It is more obvious where the renderer is.
Add new "input is ready for processing" signal.

Modify renderer to be usable with new event loop and App class.
Remove cyclic dependencies.
